requirements.txt

amqp==2.2.2
argon2-cffi==16.3.0
bcrypt==3.1.4
billiard==3.5.0.3
boto==2.48.0
boto3==1.7.5
botocore==1.10.5
celery==4.1.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11
django-annoying==0.10.4
django-appconf==1.0.2
django-bootstrap3==9.1.0
django-crispy-forms==1.7.0
django-haystack==2.6.1
django-js-asset==0.1.1
django-mathfilters==0.4.0
django-mptt==0.9.0
django-niji==0.1.0
django-photologue==3.6
django-sortedm2m==1.5.0
django-storages==1.6.6
django-tinymce==2.6.0
django-toolbelt==0.0.1
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
docutils==0.14
easy-thumbnails==2.5
ExifRead==2.1.2
gunicorn==19.7.1
idna==2.6
jmespath==0.9.3
kombu==4.1.0
markdown2==2.3.5
mistune==0.8.3
Pillow==5.0.0
psycopg2==2.7.3.2
pybbm==0.19.0
pycparser==2.18
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.3
redis==2.10.6
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
South==1.0.2
static3==0.7.0
stripe==1.79.1
Unidecode==1.0.22
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
wincertstore==0.2
xxhash==0.6.1
